+++ src/document/parsers/simples/das/schema.js
@@ -1,0 +1,10 @@
+module.exports = Object.freeze({
+  version: '1.0',
+  required: [
+    'identificacao.cnpj',
+    'identificacao.competencia',
+    'documento.tipo',
+    'datas.vencimento',
+    'valores.valorDocumento'
+  ]
+});