+++ src/document/parsers/simples/extrato.pgdas.parser.js
@@ -1,13 +1,1 @@
-exports.parse = (text) => {
-    return {
-        identificacao: {},
-        documento: {
-            tipo: 'EXTRATO_PGDAS'
-        },
-        datas: {},
-        valores: {},
-        extras: {
-            mensagem: 'Parser ainda não implementado.'
-        }
-    };
-};
+module.exports = require('./extrato');