+++ src/document/parsers/simples/extrato/schema.js
@@ -1,0 +1,8 @@
+module.exports = Object.freeze({
+  version: '1.0',
+  required: [
+    'identificacao.cnpj',
+    'identificacao.competencia',
+    'extrato'
+  ]
+});