+++ src/document/parsers/simples/relatorio/index.js
@@ -1,29 +1,9 @@
-const parseIdentificacao = require('./parse.identificacao');
-const parseResumo = require('./parse.resumo');
-const parsePartilha = require('./parse.partilha');
-const {
-    parseApuracoes,
-    consolidateByAnexo
-} = require('./parse.apuracoes');
+const parser = require('./parser');
+const schema = require('./schema');
+const rules = require('./rules');
 
-module.exports.name = 'relatorio-simples';
-
-module.exports.parse = function parseRelatorioSimples(text) {
-    const apuracoes = parseApuracoes(text);
-
-    return {
-        identificacao: parseIdentificacao(text),
-        documento: {
-            tipo: 'RELATORIO_SIMPLES'
-        },
-        resumo: parseResumo(text),
-        apuracoes,
-        consolidacaoPorAnexo: consolidateByAnexo(apuracoes),
-        partilha: parsePartilha(text),
-        extras: {
-            origem: 'DOMINIO_SISTEMAS',
-            sistemaLicenciado: /Sistema licenciado para/i.test(text),
-            quantidadeApuracoes: apuracoes.length
-        }
-    };
+module.exports = {
+  ...parser,
+  schema,
+  rules
 };