+++ src/document/parsers/simples/relatorio/schema.js
@@ -1,0 +1,9 @@
+module.exports = Object.freeze({
+  version: '1.0',
+  required: [
+    'identificacao.cnpj',
+    'identificacao.competencia',
+    'documento.tipo',
+    'resumo.valorSimplesRecolher'
+  ]
+});